Refactor list
